simplified the code
